refactor(backup): Move stop backup logic in a file
